% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Shiny Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{inputs from UI}

\item{output}{outputs to UI}

\item{session}{session to be used and observed}
}
\description{
Shiny Server
}
