% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{build_df}
\alias{build_df}
\title{Insert columns in df created above with the matches between each column of the df above and the ingredient list}
\usage{
build_df(df, ingredient_vector)
}
\arguments{
\item{df}{}

\item{ingredient_vector}{}
}
\description{
Insert columns in df created above with the matches between each column of the df above and the ingredient list
}
