% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_score.R
\name{category_health}
\alias{category_health}
\title{Sort nutrition scores into categories for Shiny output}
\usage{
category_health(marmiton_recipes)
}
\arguments{
\item{marmiton_recipes}{a dataframe to add a health index column to}
}
\value{
a dataframe with an extra column
}
\description{
Sort nutrition scores into categories for Shiny output
}
