% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{concat_matches}
\alias{concat_matches}
\title{Concat matches by unioning list columns}
\usage{
concat_matches(df, ...)
}
\arguments{
\item{df}{food dataframe from OpenFoodFacts}

\item{...}{matches column names}
}
\value{
data.frame with list of ingredient matches
}
\description{
Concat matches by unioning list columns
}
