% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{convert_to_ingredient_list}
\alias{convert_to_ingredient_list}
\title{List all the ingredients from the Marmiton recipe dataset}
\usage{
convert_to_ingredient_list(scanned_list)
}
\arguments{
\item{scanned_list}{list of products from database}
}
\value{
vector
}
\description{
List all the ingredients from the Marmiton recipe dataset
}
