% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{convert_to_ingredient_list2}
\alias{convert_to_ingredient_list2}
\title{Convert bar code to generic ingredient}
\usage{
convert_to_ingredient_list2(scanned_codes)
}
\arguments{
\item{scanned_codes}{list of bar codes from database}
}
\value{
vector
}
\description{
Convert bar code to generic ingredient
}
