% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{fetch_data_from_openfoodfacts}
\alias{fetch_data_from_openfoodfacts}
\title{Fetch Food Data from OpenFoodFacts - French dataset}
\usage{
fetch_data_from_openfoodfacts()
}
\description{
Fetch Food Data from OpenFoodFacts - French dataset
}
