% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmitoncrawler.R
\name{fetch_recipe_detail}
\alias{fetch_recipe_detail}
\title{Fetch recipe details}
\usage{
fetch_recipe_detail(url)
}
\arguments{
\item{url}{character}
}
\value{
data.frame
}
\description{
Fetch recipe details
}
