% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmitoncrawler.R
\name{fetch_top_100_recipes}
\alias{fetch_top_100_recipes}
\title{Fetch Top 100 recipes from Marmiton website}
\usage{
fetch_top_100_recipes()
}
\description{
Fetch Top 100 recipes from Marmiton website
}
