% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{filter_categories}
\alias{filter_categories}
\title{Get the best category of the product}
\usage{
filter_categories(food_dataset)
}
\arguments{
\item{food_dataset}{data.frame from OpenFoodFacts}
}
\value{
filtered categories dataset
}
\description{
Get the best category of the product
}
