% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmitoncrawler.R
\name{gather_recipes_to_csv}
\alias{gather_recipes_to_csv}
\title{Write the recies data into a csv file}
\usage{
gather_recipes_to_csv(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
character
}
\description{
Write the recies data into a csv file
}
