% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missing_scores.R
\name{generate_missing_scores}
\alias{generate_missing_scores}
\title{Algorithm in two steps:
first step complete missing values on the continuous predictor variables with 
an iterative PCA from missMDA package
second step is a linear regression trained on the dataset with the non missing nutrition score
and used to predict the missing ones}
\usage{
generate_missing_scores()
}
\value{
a dataframe with barcode and corresponding grades
}
\description{
Algorithm in two steps:
first step complete missing values on the continuous predictor variables with 
an iterative PCA from missMDA package
second step is a linear regression trained on the dataset with the non missing nutrition score
and used to predict the missing ones
}
