% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get.matrix.subset.containance}
\alias{get.matrix.subset.containance}
\title{Calcul matrix of containance between different sets: for a given set i, m[i,j]=1 if the set j contains the set i}
\usage{
get.matrix.subset.containance(all.subset)
}
\arguments{
\item{all.subset}{}
}
\value{
matrix of containance
}
\description{
Calcul matrix of containance between different sets: for a given set i, m[i,j]=1 if the set j contains the set i
}
