% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{get_all_product_names}
\alias{get_all_product_names}
\title{List all the product names from the Open foods dataset}
\usage{
get_all_product_names()
}
\value{
vector
}
\description{
List all the product names from the Open foods dataset
}
