% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_combinations}
\alias{get_best_combinations}
\title{Returns the subsets that have one or more appearances in the list of marmiton's ingredients}
\usage{
get_best_combinations(list)
}
\arguments{
\item{list}{list of subsets of ingredients}
}
\value{
list of subsets usable in some recipes
}
\description{
Returns the subsets that have one or more appearances in the list of marmiton's ingredients
}
