% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_recipes}
\alias{get_best_recipes}
\title{Give the list of all the recipes given by the three functions called, depending on some criteria}
\usage{
get_best_recipes(ingredients_to_use, proportion_of_recipe = 30,
  minimum_ingredients_to_use = 0, must_include = NULL)
}
\arguments{
\item{ingredients_to_use}{list of ingredients to use (in our fridge for example)}

\item{proportion_of_recipe}{proportion of the recipe covered by the subset of the ingredients}

\item{minimum_ingredients_to_use}{minimum number of ingredients to really use within the ingredients_to_use}

\item{must_include}{ingredients to be included}
}
\value{
list of recipes
}
\description{
Give the list of all the recipes given by the three functions called, depending on some criteria
}
