% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_recipes_du_chef}
\alias{get_best_recipes_du_chef}
\title{Calculate the best combination of ingredients
i.e the one that appears the most in the list of unique, with respect to the parameters
and returns a list of recipies containing this combination.
Test with get_best_recipes_du_chef(c("carotte", "jambon", "fromage"), 20, 1, T)}
\usage{
get_best_recipes_du_chef(ingredients_to_use, proportion_of_recipe,
  minimum_ingredients_to_use, must_include)
}
\arguments{
\item{ingredients_to_use}{Set of ingredients}

\item{proportion_of_recipe}{}

\item{minimum_ingredients_to_use}{}

\item{must_include}{}
}
\value{
List of recipes
}
\description{
Calculate the best combination of ingredients
i.e the one that appears the most in the list of unique, with respect to the parameters
and returns a list of recipies containing this combination.
Test with get_best_recipes_du_chef(c("carotte", "jambon", "fromage"), 20, 1, T)
}
