% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_recipies}
\alias{get_best_recipies}
\title{Give the list of all the recipes given by the three functions called}
\usage{
get_best_recipies(ingredients_to_use, proportion_of_recipe = 30,
  minimum_ingredients_to_use = 0, must_include = "nothing_bro")
}
\arguments{
\item{ingredients_to_use}{}

\item{proportion_of_recipe}{}

\item{minimum_ingredients_to_use}{}

\item{must_include}{}
}
\description{
Give the list of all the recipes given by the three functions called
}
