% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{get_last_category}
\alias{get_last_category}
\title{Split categories and get the last item listed}
\usage{
get_last_category(categories)
}
\arguments{
\item{categories}{character with multiple categories separated by a comma}
}
\value{
category with a single element
}
\description{
Split categories and get the last item listed
}
