% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_nb_best_combinations}
\alias{get_nb_best_combinations}
\title{returns the number of appearance of each interresting subset in the list of unique, 
with taking into account - this time - the percentage of ingredient that cover the recepies 
(e.g. if recipie is carotte oeuf fromage, the subset "carotte" covers 33\% of the recipie)
if percentage demanded is 40\% than the subset "carotte" won't appear.}
\usage{
get_nb_best_combinations(all_subsets, percentage)
}
\arguments{
\item{all_subsets}{all possible subsets from the original list of ingredients}

\item{percentage}{miminum percentage of the recipe to be covered by the subset of ingredients}
}
\value{
list of occurences for each subset
}
\description{
returns the number of appearance of each interresting subset in the list of unique, 
with taking into account - this time - the percentage of ingredient that cover the recepies 
(e.g. if recipie is carotte oeuf fromage, the subset "carotte" covers 33\% of the recipie)
if percentage demanded is 40\% than the subset "carotte" won't appear.
}
