% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_number_for_one_combination}
\alias{get_number_for_one_combination}
\title{Count the number of appearances of one combination of ingredients in the list of marmiton recipes}
\usage{
get_number_for_one_combination(one_combination, percentage)
}
\arguments{
\item{one_combination}{one particular combination in all the subset}

\item{percentage}{a percentage}
}
\description{
Count the number of appearances of one combination of ingredients in the list of marmiton recipes
}
