% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_number_of_each_combinations}
\alias{get_number_of_each_combinations}
\title{Give the number of appearances of each subset in the list of unique}
\usage{
get_number_of_each_combinations(set, minimum_to_use, must_include = NULL)
}
\arguments{
\item{set}{the set of ingredient you have}

\item{minimum_to_use}{in the recipe}

\item{must_include}{ingredient}
}
\description{
Give the number of appearances of each subset in the list of unique
}
