% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_position_of_future_empty}
\alias{get_position_of_future_empty}
\title{Enables to iterate only on the subset that don't contain a subset with no appearances in the list of unique
That is why we set percentage to 0 in the is_contained function.
Indeed if it wasn't, we would not iterate on set that contains a subset with no appearance in the list of unique.
Yet having no appearances in that list can mean, that the percentage rule was not passed,
and not because it doesn't appear.}
\usage{
get_position_of_future_empty(matrix, i)
}
\arguments{
\item{matrix}{}

\item{i}{}
}
\description{
Enables to iterate only on the subset that don't contain a subset with no appearances in the list of unique
That is why we set percentage to 0 in the is_contained function.
Indeed if it wasn't, we would not iterate on set that contains a subset with no appearance in the list of unique.
Yet having no appearances in that list can mean, that the percentage rule was not passed,
and not because it doesn't appear.
}
