% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_receipe_with_least_ingredients_to_add}
\alias{get_receipe_with_least_ingredients_to_add}
\title{Calculate the combination that is the closest to the combination of one receipe
and returns the corresponding receipe}
\usage{
get_receipe_with_least_ingredients_to_add(ingredients_to_use,
  minimum_ingredients_to_use, must_include)
}
\arguments{
\item{ingredients_to_use}{}

\item{minimum_ingredients_to_use}{}

\item{must_include}{}
}
\value{
list of recipes
}
\description{
Calculate the combination that is the closest to the combination of one receipe
and returns the corresponding receipe
}
