% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_recipe_with_least_ingredients_to_add}
\alias{get_recipe_with_least_ingredients_to_add}
\title{Calculate the combination that is the closest to the combination of one recipe
and returns the corresponding recipe}
\usage{
get_recipe_with_least_ingredients_to_add(ingredients_to_use,
  minimum_ingredients_to_use, must_include)
}
\arguments{
\item{ingredients_to_use}{in the recipe}

\item{minimum_ingredients_to_use}{in the recipe}

\item{must_include}{ingredient in the recipe}
}
\value{
list of recipes
}
\description{
Calculate the combination that is the closest to the combination of one recipe
and returns the corresponding recipe
}
