% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_subsets_from}
\alias{get_subsets_from}
\title{Quick way to obtain all the subsets given a set}
\usage{
get_subsets_from(set)
}
\arguments{
\item{set}{}
}
\value{
list of subsets
}
\description{
Quick way to obtain all the subsets given a set
}
