% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmitoncrawler.R
\name{get_top_100_links}
\alias{get_top_100_links}
\title{Return Top 100 recipes links of Marmiton website}
\usage{
get_top_100_links()
}
\value{
data.frame
}
\description{
Return Top 100 recipes links of Marmiton website
}
