% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{ingredients_list}
\alias{ingredients_list}
\title{List all the ingredients from the Marmiton recipe dataset}
\usage{
ingredients_list()
}
\value{
vector
}
\description{
List all the ingredients from the Marmiton recipe dataset
}
