% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{is_contained}
\alias{is_contained}
\title{The function that will look if something is contained in something
if number of ingredients if too far (in percentage)
from the number of ingredients in the recipe, is_contained don't count the containance.
The percentage parameter is set to 0 by default for reasons you will understand after}
\usage{
is_contained(v1, v2, percentage = 0)
}
\arguments{
\item{v1}{vector to be included in v2}

\item{v2}{vector to contain v1}

\item{percentage}{}
}
\value{
boolean
}
\description{
The function that will look if something is contained in something
if number of ingredients if too far (in percentage)
from the number of ingredients in the recipe, is_contained don't count the containance.
The percentage parameter is set to 0 by default for reasons you will understand after
}
