% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match_list_column_to_ingredients}
\alias{match_list_column_to_ingredients}
\title{Match list_column words with marmiton ingredients list
Create a matches_column}
\usage{
match_list_column_to_ingredients(df, ingredient_vector, list_column,
  matches_column)
}
\arguments{
\item{df}{food dataframe}

\item{ingredient_vector}{ingredient list}

\item{list_column}{}

\item{matches_column}{}
}
\value{
data.frame
}
\description{
Match list_column words with marmiton ingredients list
Create a matches_column
}
