% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{match_open_food_data_to_marmiton_ingredients}
\alias{match_open_food_data_to_marmiton_ingredients}
\title{Creates a matching column for the OpenFoodData set with Marmiton generic ingredients}
\usage{
match_open_food_data_to_marmiton_ingredients(df)
}
\arguments{
\item{df}{food dataset from OpenFoodData}
}
\value{
data.frame
}
\description{
Creates a matching column for the OpenFoodData set with Marmiton generic ingredients
}
