% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{prepare_df}
\alias{prepare_df}
\title{Create a dataframe that splits up the product name}
\usage{
prepare_df(df, col_name)
}
\arguments{
\item{df}{}

\item{col}{}
}
\value{
data.frame
}
\description{
Create a dataframe that splits up the product name
}
