% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{preprocess_set}
\alias{preprocess_set}
\title{Give all the subsets of the set according to the conditions wanted (minimum ingredient to use, must include ...)}
\usage{
preprocess_set(set, minimum_to_use, must_include = NULL)
}
\arguments{
\item{set}{}

\item{minimum_to_use}{a minimum of ingredient you want to use}

\item{must_include}{the ingredient that must be in the recipe}
}
\value{
subsets arranged in increasing length
}
\description{
Give all the subsets of the set according to the conditions wanted (minimum ingredient to use, must include ...)
}
