% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_score.R
\name{recipe_score}
\alias{recipe_score}
\title{Takes a list of ingredients and returns the average nutrition grade}
\usage{
recipe_score(list_of)
}
\arguments{
\item{list_of}{list of ingredients in a recipe}
}
\value{
returns the average nutrition score for the list of ingredients (of a recipe)
}
\description{
Takes a list of ingredients and returns the average nutrition grade
}
