% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{remove_dash}
\alias{remove_dash}
\title{Change dash to space}
\usage{
remove_dash(x)
}
\arguments{
\item{x}{a string of character}
}
\value{
string without dashes but spaces
}
\description{
Change dash to space
}
\examples{
remove_dash('abc-def')
}
