% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{remove_dashes_for_marmiton}
\alias{remove_dashes_for_marmiton}
\title{Remove all dashes from marmiton_recipe_ingredients dataframe}
\usage{
remove_dashes_for_marmiton()
}
\value{
marmiton recipes without dashes
}
\description{
Remove all dashes from marmiton_recipe_ingredients dataframe
}
