% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{remove_ingredients_you_always_have}
\alias{remove_ingredients_you_always_have}
\title{Remove common ingredients such as salt, pepper, etc.}
\usage{
remove_ingredients_you_always_have(list)
}
\arguments{
\item{list}{list of ingredients}
}
\value{
list of filtered ingredients
}
\description{
Remove common ingredients such as salt, pepper, etc.
}
