% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{remove_tiret_for_marmiton}
\alias{remove_tiret_for_marmiton}
\title{a list of list of character}
\usage{
remove_tiret_for_marmiton()
}
\arguments{
\item{list}{with all the recipes}
}
\description{
a list of list of character
}
