% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{replace}
\alias{replace}
\title{Decision rule to choose which ingredient to use to replace the product name}
\usage{
replace()
}
\description{
Decision rule to choose which ingredient to use to replace the product name
}
