% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_score.R
\name{score_column}
\alias{score_column}
\title{Creates a column in the marmiton recipes dataframe with the average overall nutritional score for all ingredients used.}
\usage{
score_column(marmiton_recipes)
}
\arguments{
\item{marmiton_recipes}{a dataframe with the information about each recipe}
}
\value{
returns the dataframe used as an input with an extra column
}
\description{
Creates a column in the marmiton recipes dataframe with the average overall nutritional score for all ingredients used.
}
