% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_sorting.R
\name{sort_budget}
\alias{sort_budget}
\title{Rank by Cheapest recipes}
\usage{
sort_budget(recipes)
}
\arguments{
\item{recipes}{output tibble of recipes in the shiny app}
}
\value{
tibble of recipes sorted according to their cost
}
\description{
Rank by Cheapest recipes
}
