% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_sorting.R
\name{sort_difficulty}
\alias{sort_difficulty}
\title{Easiest recipes}
\usage{
sort_difficulty(recipes)
}
\arguments{
\item{recipes}{tibble of recipes}
}
\value{
tibble of recipes sorted according to the budget
}
\description{
Easiest recipes
}
