% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_sorting.R
\name{sort_time}
\alias{sort_time}
\title{Rank by Quickest recipes}
\usage{
sort_time(recipes)
}
\arguments{
\item{recipes}{output tibble of recipes in the shiny app}
}
\value{
tibble of recipes sorted according to the total time required for the recipe
}
\description{
Rank by Quickest recipes
}
