% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{split_column_words_to_new_col}
\alias{split_column_words_to_new_col}
\title{Create a dataframe that splits up the words of column_name to words_column_name}
\usage{
split_column_words_to_new_col(df, column_name, words_column_name)
}
\arguments{
\item{df}{df to apply the split on}

\item{column_name}{column to split}

\item{words_column_name}{name of new column}
}
\value{
data.frame
}
\description{
Create a dataframe that splits up the words of column_name to words_column_name
}
