% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{split_columns_to_match}
\alias{split_columns_to_match}
\title{Splits mutliple columns into list columns for possible matching}
\usage{
split_columns_to_match(df)
}
\arguments{
\item{df}{data.frame with strings}
}
\value{
data.frame
}
\description{
Splits mutliple columns into list columns for possible matching
}
