% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{convert_to_ingredient_list}
\alias{convert_to_ingredient_list}
\title{Convert product name to generic ingredient}
\usage{
convert_to_ingredient_list(scanned_list)
}
\arguments{
\item{scanned_list}{list of products from database}
}
\value{
vector of generic ingredients
}
\description{
Convert product name to generic ingredient
}
