% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_ingredients.R, R/ingredients.R
\name{convert_to_ingredient_list}
\alias{convert_to_ingredient_list}
\title{Convert product names to ingredient lists}
\usage{
convert_to_ingredient_list(scanned_list)

convert_to_ingredient_list(scanned_list)
}
\arguments{
\item{scanned_list}{}

\item{scanned_list}{list of products from database}
}
\value{
list

vector
}
\description{
Convert product names to ingredient lists

Convert product name to generic ingredient
}
