% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marmitoncrawler.R
\name{fetch_recipe_detail}
\alias{fetch_recipe_detail}
\title{Fetch fromatted recipe ingredient list and preparation instruction steps}
\usage{
fetch_recipe_detail(url)
}
\arguments{
\item{url}{character}
}
\value{
data.frame
}
\description{
Fetch fromatted recipe ingredient list and preparation instruction steps
}
