% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_output.R
\name{function_div}
\alias{function_div}
\title{Recipes output in HTML for the Shiny App}
\usage{
function_div(URL, recipeTitle, pic_URL, recipeDifficulty, recipeCost,
  recipePreparationTime, recipeCookingTime)
}
\arguments{
\item{URL}{}

\item{recipeTitle}{}

\item{recipeDifficulty}{}

\item{recipePreparationTime}{}

\item{recipeCookingTime}{}
}
\description{
Recipes output in HTML for the Shiny App
}
