% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_combinations}
\alias{get_best_combinations}
\title{returns the subsets that have one or more appearance in the list of unique}
\usage{
get_best_combinations(list)
}
\arguments{
\item{list}{}
}
\description{
returns the subsets that have one or more appearance in the list of unique
}
