% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_best_recipe_using_most_ingredients}
\alias{get_best_recipe_using_most_ingredients}
\title{Calculate the biggest combination of ingredient that appears at least one time
and returns the corresponding receipies}
\usage{
get_best_recipe_using_most_ingredients(ingredients_to_use, must_include)
}
\arguments{
\item{ingredients_to_use}{}

\item{must_include}{}
}
\value{
list of recipes
}
\description{
Calculate the biggest combination of ingredient that appears at least one time
and returns the corresponding receipies
}
