% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{get_last_category}
\alias{get_last_category}
\title{Split and get the last one}
\usage{
get_last_category(categories)
}
\arguments{
\item{categories}{character with many categories separated by comma}
}
\value{
category
}
\description{
Split and get the last one
}
