% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_number_of_each_combinations}
\alias{get_number_of_each_combinations}
\title{Give the number of appearances of each subset in the list of unique}
\usage{
get_number_of_each_combinations(set, minimum_to_use, must_include)
}
\arguments{
\item{set}{}

\item{minimum_to_use}{}

\item{must_include}{}
}
\description{
Give the number of appearances of each subset in the list of unique
}
