% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_score.R
\name{nutrition_score}
\alias{nutrition_score}
\title{Get the average nutrition score of an ingredient}
\usage{
nutrition_score(ingredient)
}
\arguments{
\item{ingredient}{an ingredient}
}
\value{
an average nutrition score for the ingredient input
}
\description{
Get the average nutrition score of an ingredient
}
