% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_output.R
\name{recipe_output}
\alias{recipe_output}
\title{Get a list of html tags, iterated over an entire dataframe}
\usage{
recipe_output(tibble)
}
\arguments{
\item{tibble}{}
}
\value{
html tag
}
\description{
Get a list of html tags, iterated over an entire dataframe
}
