% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_sorting.R
\name{sort_time}
\alias{sort_time}
\title{Quickest recipes}
\usage{
sort_time(recipes)
}
\arguments{
\item{recipes}{tibble of recipes}
}
\value{
tibble of recipes sorted according to the total time
}
\description{
Quickest recipes
}
